Refactor connection and request handling in BackendForFE
This commit refactors the connection and request handling in the BackendForFE struct. It replaces the use of the `wasabi.Connection` type with a new `ConnState` struct, which contains the connection and a current ID. The `connection` map is updated to use the new `ConnState` struct, and the `requests` map is changed to use `int64` as the key type. Additionally, the `Handle` method now uses the `ConnState` struct to handle requests.

The changes in this commit improve the organization and clarity of the code, making it easier to manage connections and requests in the BackendForFE struct.

+++ pkg/handlers/handler.go
@@ -32,7 +32,7 @@ type RequestRunConfig struct {
 
 type TemplateData struct {
 	Params map[string]any
-	ReqID  int
+	ReqID  int64
 }
 
 func NewCallHandler(config *HandlersConfig) (*CallHandler, error) {
@@ -108,7 +108,7 @@ type RequesIter struct {
 	mu        sync.Mutex
 }
 
-func (r *RequesIter) Next(ctx context.Context, id int) ([]byte, chan []byte, error) {
+func (r *RequesIter) Next(ctx context.Context, id int64) ([]byte, chan []byte, error) {
 	if r.pos >= len(r.reqs) {
 		return nil, nil, ErrIterDone
 	}
